add new IdleEventHandler for Telnet Communication in QoS
